remove Game from front-end/src/index.js
