+++ front-end/src/index.js
@@ -1,14 +1,14 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import io from 'socket.io-client';
-import { Game } from './features/tic-tac-toe';
+import App from './features/app';
 
 import './index.css';
 
 const socket = io('http://localhost:8090');
 socket.on('state-change', (state) =>
   ReactDOM.render(
-    <Game dispatch={(action) => socket.emit('action', action)} gameState={state} />,
+    <App dispatch={(action) => socket.emit('action', action)} state={state} />,
     document.getElementById('root'),
   ),
 );
